requirements.txt
PyQt5==5.15.4
paho-mqtt
playsound
imutils
pandas
PyOpenGL==3.1.0
joblib
scipy
tqdm
numpy
numba
filterpy
yacs
smplx
trimesh
pyrender
scikit-image
h5py
pytube
scikit-learn
gdown
PyYAML
pillow
progress
matplotlib
tensorflow
tensorboard
torch
torchvision
torchaudio
scikit-video
llvmlite
opencv-python==4.2.0.34
ninja
markupsafe
PyGObject
pycairo
PyTurboJPEG
shapely
scikit-spatial
python-prctl
vidgear[core]==0.2.6
multipledispatch
seaborn
pybind11
mypy # for generating stubs (pybind11)
docker
moviepy
pyvistaqt
psutil
pyqtgraph
